Got all compiled. Failing linking
git-svn-id: https://roadrunnerlib.googlecode.com/svn/trunk@121 c0027639-9190-0b57-01eb-f09a03790dae

+++ source/rrRoadRunner.cpp
@@ -2626,7 +2626,7 @@ string RoadRunner::writeSBML()
     }
 
     array = getGlobalParameterIds();
-    for (int i = 0; i < min(array.Count(), state.mGlobalParameters.size()); i++)
+    for (int i = 0; i < min((int) array.Count(), (int) state.mGlobalParameters.size()); i++)
     {
         NOM.setValue((string)array[i], state.mGlobalParameters[i]);
     }
